:truck: 엔트리 section 관련 이벤트 entry-section.js 로 통합 #217
